+++ comp2013/src/main/java/comp2013/IView.java
@@ -29,7 +29,7 @@ public interface IView {
      * Sets the background image of the game.
      * @param imageView ImageView that holds the image.
      */
-    public void setBackgroundImage(ImageView imageView);
+    void setBackgroundImage(ImageView imageView, String imageID);
 
     /**
      * Draws the current score of the game.
@@ -45,4 +45,9 @@ public interface IView {
      * Handles what happens to the view when a game over condition is met.
      */
     void gameOverScreen();
+
+    /**
+     * Restarts the game by resetting the UI.
+     */
+    void restartGame();
 }